GCP/bankmarketing/requirements.txt
#Django
django>=1.9.12
djangorestframework>=3.7.7
django-cors-headers>=2.2.0
drf-yasg>=1.7.0
whitenoise==3.2.2

numpy>=1
#cryptography==1.8.1
cffi>=1.11.4
pandas>=0.20.1
scipy>=0.19.0
scikit-learn>=0.19.0

